Add serializeJSON member function to Array, Query, Struct
Member functions with the first arg of type 'any' do not appear to work out of the box, changing the parsing system seems like it would have other implications so I did not take that approach, but it might be superior, since you can technically do a boolean, string, numeric, etc. serializeJSON call, but it's pointless.